optimization
move the ValidatorStatus and the GetStatus inside the sidechain helper;
remove the vesting flag and use only the vesting-period;
update the readme;
