+++ command/sidechain/helper.go
@@ -9,7 +9,6 @@ import (
 	"github.com/0xPolygon/polygon-edge/command/polybftsecrets"
 	"github.com/0xPolygon/polygon-edge/consensus/polybft"
 	"github.com/0xPolygon/polygon-edge/consensus/polybft/contractsapi"
-	"github.com/0xPolygon/polygon-edge/consensus/polybft/validator"
 	"github.com/0xPolygon/polygon-edge/consensus/polybft/wallet"
 	"github.com/0xPolygon/polygon-edge/contracts"
 	"github.com/0xPolygon/polygon-edge/helper/hex"
@@ -83,6 +82,30 @@ func CreateTransaction(sender ethgo.Address, to *ethgo.Address, input []byte, va
 	return txn
 }
 
+// Define the ValidatorStatus enum
+type ValidatorStatus int
+
+const (
+	None ValidatorStatus = iota
+	Registered
+	Active
+	Banned
+)
+
+// Function to get ValidatorStatus based on number value
+func GetStatus(value int) ValidatorStatus {
+	switch value {
+	case 1:
+		return Active
+	case 2:
+		return Active
+	case 3:
+		return Banned
+	default:
+		return None
+	}
+}
+
 // GetValidatorInfo queries HydraChain smart contract to retrieve the validator info for given address
 func GetValidatorInfo(txRelayer txrelayer.TxRelayer, validatorAddr ethgo.Address) (*polybft.ValidatorInfo, error) {
 	var getValidatorFn = &contractsapi.GetValidatorHydraChainFn{
@@ -135,7 +158,7 @@ func GetValidatorInfo(txRelayer txrelayer.TxRelayer, validatorAddr ethgo.Address
 		Address:             validatorAddr,
 		Stake:               stake,
 		WithdrawableRewards: withdrawableRewards,
-		IsActive:            validator.Active == validator.GetStatus(int(status)),
+		IsActive:            Active == GetStatus(int(status)),
 	}
 
 	return validatorInfo, nil
